pyproject.toml
[tool.poetry]
name = "ebtc-multisig"
version = "0.1.0"
description = "eBTC's EVM multisig operations."
authors = ["Gosuto Inzasheru <mail@gosuto.ai>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
brownie-safe = "*"
dotmap = "*"
eth-brownie = "1.20.5"
pandas = "*"
pycoingecko = "^3.1.0"
rich = "*"
tabulate = "*"
tqdm = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
